Feat(be): `Reservation` 연관관계 수정
관리자, 참가자 연관관계 별도 지정
