--- backend/grabtable/src/main/java/edu/skku/grabtable/domain/Reservation.java
@@ -1,22 +1,0 @@
-package edu.skku.grabtable.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-
-@Entity
-@Getter
-public class Reservation extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Integer numberOfUser;
-
-    @Enumerated(EnumType.STRING)
-    private ReservationStatus status;
-}
