--- backend/grabtable/src/main/java/edu/skku/grabtable/domain/ReservationStatus.java
@@ -1,5 +1,0 @@
-package edu.skku.grabtable.domain;
-
-public enum ReservationStatus {
-    VALID, PENDING, END
-}
