+++ backend/grabtable/src/main/java/edu/skku/grabtable/reservation/domain/ReservationStatus.java
@@ -1,0 +1,5 @@
+package edu.skku.grabtable.reservation.domain;
+
+public enum ReservationStatus {
+    INVALID, PENDING, CONFIRMED
+}
